lsp: Extract some functionality from native preview
... so that this can get reused for the WASM flavor.
